Extract functions and types from ContactSensor capability
Related to #44
