+++ src/hubitat-capabilities/index.ts
@@ -1,6 +1,7 @@
 /* FUNCTIONS */
 
 export * from './functions/get-alarm-setting.function';
+export * from './functions/get-contact-sensor-status.function';
 export * from './functions/get-cooling-setpoint.function';
 export * from './functions/get-heating-setpoint.function';
 export * from './functions/get-supported-thermostat-fan-modes.function';
@@ -10,6 +11,8 @@ export * from './functions/get-thermostat-mode.function';
 export * from './functions/get-thermostat-operating-state.function';
 export * from './functions/get-thermostat-schedule.function';
 export * from './functions/get-thermostat-setpoint.function';
+
+export * from './functions/is-open.function';
 
 export * from './functions/set-alarm-setting.function';
 export * from './functions/set-cooling-setpoint.function';
@@ -29,6 +32,7 @@ export * from './functions/stop-level-change.function';
 
 export * from './types/alarm-setting.type';
 export * from './types/level-change-direction.type';
+export * from './types/open-status.type';
 export * from './types/thermostat-fan-modes.type';
 export * from './types/thermostat-modes.type';
 export * from './types/thermostat-operating-modes.type';
