+++ src/hubitat-capabilities/types/open-status.type.ts
@@ -1,0 +1,9 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+/**
+ * Possible open-closed status.
+ */
+export type EOpenStatus = 'closed' | 'open';
